requirements.txt
langchain-google-genai==0.0.12
langchain-community==0.0.38
faiss-cpu==1.8.0
python-dotenv==1.0.1
fastapi==0.115.2
uvicorn==0.32.0